Refactor debug toggles across modules
Introduce centralized debug configuration flags (`debug::config`) and add per-library debug wrapper headers (Art-Net, board, display, DMX, network, OSC, RDM, showfile, etc.) so tracing can be enabled per subsystem without forcing `NDEBUG` changes. Update many sources to use the new scoped debug macros, switch stack/vector gating to `CONFIG_DEBUG_STACK`, and clean up related logging/type formatting and small correctness/style issues discovered during the migration.

+++ lib-device/src/sc16is740.cpp
@@ -150,9 +150,9 @@ void SC16IS740::SetBaud(uint32_t baud) {
     WriteRegister(SC16IS7X0_DLH, static_cast<uint8_t>((kDivisor >> 8) & 0xFF), false);
     WriteRegister(SC16IS7X0_LCR, kRegisterLcr, false);
 
-    DEBUG_PRINTF("prescaler=%u", prescaler);
-    DEBUG_PRINTF("kDivisor=%u", kDivisor);
-    DEBUG_PRINTF("on_board_crystal_hz_=%u", on_board_crystal_hz_);
+    DEBUG_PRINTF("prescaler=%u", static_cast<unsigned>(prescaler));
+    DEBUG_PRINTF("kDivisor=%u", static_cast<unsigned>(kDivisor));
+    DEBUG_PRINTF("on_board_crystal_hz_=%u", static_cast<unsigned>(on_board_crystal_hz_));
 
     DEBUG_PRINTF("LCR=%.2x:%.2x", ReadRegister(SC16IS7X0_LCR, false), kRegisterLcr);
 }
